refactor: clean engine api package
