+++ pkg/engine/api/policyspec.go
@@ -1,0 +1,7 @@
+package api
+
+// PolicySpec policy
+type PolicySpec struct {
+	Name      string
+	Namespace string
+}
