+++ pkg/engine/mutate/patch/strategicMergePatch.go
@@ -23,9 +23,9 @@ func ProcessStrategicMergePatch(ruleName string, overlay interface{}, resource u
 	resp.Type = engineapi.Mutation
 
 	defer func() {
-		resp.RuleStats.ProcessingTime = time.Since(startTime)
-		resp.RuleStats.RuleExecutionTimestamp = startTime.Unix()
-		logger.V(4).Info("finished applying strategicMerge patch", "processingTime", resp.RuleStats.ProcessingTime.String())
+		resp.ExecutionStats.ProcessingTime = time.Since(startTime)
+		resp.ExecutionStats.Timestamp = startTime.Unix()
+		logger.V(4).Info("finished applying strategicMerge patch", "processingTime", resp.ExecutionStats.ProcessingTime.String())
 	}()
 
 	overlayBytes, err := json.Marshal(overlay)
